+++ durabletask-azuremanaged/durabletask/azuremanaged/preview/on_demand_sandbox/worker.py
@@ -9,13 +9,15 @@ from typing import Iterator, Optional
 
 from azure.identity import ManagedIdentityCredential
 
-from durabletask.azuremanaged.preview.on_demand_sandbox.client import (
+from durabletask.azuremanaged.preview.on_demand_sandbox.declarations import (
     DEFAULT_MAX_CONCURRENT_ACTIVITIES,
     DEFAULT_WORKER_PROFILE_ID,
-    _OnDemandSandboxActivitiesGrpcClient,
     build_on_demand_sandbox_worker_heartbeat,
     build_on_demand_sandbox_worker_start,
     resolve_activity_names,
+)
+from durabletask.azuremanaged.preview.on_demand_sandbox.transport import (
+    OnDemandSandboxActivitiesGrpcTransport,
 )
 from durabletask.azuremanaged.worker import DurableTaskSchedulerWorker
 import durabletask.internal.shared as shared
@@ -118,7 +120,7 @@ class OnDemandSandboxWorker(DurableTaskSchedulerWorker):
         retry_delay = 1.0
         while not self._on_demand_sandbox_registration_stop.is_set():
             try:
-                client = _OnDemandSandboxActivitiesGrpcClient(
+                client = OnDemandSandboxActivitiesGrpcTransport(
                     host_address=self._on_demand_sandbox_host_address,
                     taskhub=self._on_demand_sandbox_taskhub,
                     token_credential=self._on_demand_sandbox_token_credential,
